# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoundApp OpenApi Documentation",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "francesco.simeoli@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://roundapp.local/api",
            "description": "RoundApp API Server"
        },
        {
            "url": "https://roundapp.local"
        }
    ],
    "paths": {
...
        "/consumers/me": {
            "get": {
                "tags": [
                    "Consumers"
                ],
                "summary": "Get Consumer",
                "description": "Get Consumer",
                "operationId": "showConsumer",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consumer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Consumers"
                ],
                "summary": "Update Consumer",
                "description": "Update Consumer",
                "operationId": "updateConsumers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consumer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/producers": {
            "get": {
                "tags": [
                    "Producers"
                ],
                "summary": "Get Producers list",
                "description": "Get Producers list",
                "operationId": "indexProducers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Producers"
                ],
                "summary": "Create Producer",
                "description": "Create Producer",
                "operationId": "storeProducers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/producers/{id}": {
            "get": {
                "tags": [
                    "Producers"
                ],
                "summary": "Get Producer",
                "description": "Get Producer",
                "operationId": "showProducer",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Producers"
                ],
                "summary": "Update Producer",
                "description": "Update Producer",
                "operationId": "updateProducers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Producers"
                ],
                "summary": "Delete Producer",
                "description": "Delete Producer",
                "operationId": "deleteProducers",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Consumer": {
                "description": "Entity Consumer",
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "surname",
                    "birth_year",
                    "gender"
                ],
                "properties": {
                    "id": {
                        "schema": "Consumer",
                        "description": "id",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "user_id": {
                        "schema": "Consumer",
                        "description": "user_id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "name": {
                        "schema": "Consumer",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "surname": {
                        "schema": "Consumer",
                        "description": "surname",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "birth_year": {
                        "schema": "Consumer",
                        "description": "birth_year",
                        "type": "integer",
                        "format": "integer"
                    },
                    "gender": {
                        "schema": "Consumer",
                        "description": "gender",
                        "type": "integer",
                        "format": "smallint"
                    },
                    "created_at": {
                        "schema": "Consumer",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "schema": "Consumer",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "schema": "Consumer",
                        "description": "deleted_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "schema": "Consumer",
                        "description": "User",
                        "type": "object",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
            "File": {
                "description": "Entity File",
                "required": [
                    "id",
                    "uuid",
                    "user_id",
                    "type",
                    "path",
                    "size",
                    "mime_type",
                    "original_name",
                    "name",
                    "description",
                    "hash_alg",
                    "hash_file"
                ],
                "properties": {
                    "id": {
                        "schema": "File",
                        "description": "id",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "uuid": {
                        "schema": "File",
                        "description": "uuid",
                        "type": "string"
                    },
                    "url": {
                        "schema": "File",
                        "description": "url",
                        "type": "string"
                    },
                    "user_id": {
                        "schema": "File",
                        "description": "user_id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "type": {
                        "schema": "File",
                        "description": "type",
                        "type": "string",
                        "format": "string",
                        "maxLength": 30
                    },
                    "path": {
                        "schema": "File",
                        "description": "path",
                        "type": "string",
                        "format": "string",
                        "maxLength": 2000
                    },
                    "fullPath": {
                        "schema": "File",
                        "description": "fullPath",
                        "type": "string",
                        "format": "string"
                    },
                    "size": {
                        "schema": "File",
                        "description": "size",
                        "type": "integer",
                        "format": "integer"
                    },
                    "mime_type": {
                        "schema": "File",
                        "description": "mime_type",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "original_name": {
                        "schema": "File",
                        "description": "original_name",
                        "type": "string",
                        "format": "string",
                        "maxLength": 300
                    },
                    "name": {
                        "schema": "File",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "maxLength": 300
                    },
                    "description": {
                        "schema": "File",
                        "description": "description",
                        "type": "string",
                        "format": "text"
                    },
                    "hash_alg": {
                        "schema": "File",
                        "description": "hash_alg",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "hash_file": {
                        "schema": "File",
                        "description": "hash_file",
                        "type": "string",
                        "format": "string",
                        "maxLength": 300
                    },
                    "created_at": {
                        "schema": "File",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "schema": "File",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "schema": "File",
                        "description": "deleted_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "schema": "File",
                        "description": "User",
                        "type": "object",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
...
            "Producer": {
                "description": "Entity Producer",
                "required": [
                    "id",
                    "user_id",
                    "image_id",
                    "name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "schema": "Producer",
                        "description": "id",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "user_id": {
                        "schema": "Producer",
                        "description": "user_id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "image_id": {
                        "schema": "Producer",
                        "description": "image_id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "name": {
                        "schema": "Producer",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "description": {
                        "schema": "Producer",
                        "description": "description",
                        "type": "string",
                        "format": "text"
                    },
                    "website": {
                        "schema": "Producer",
                        "description": "website",
                        "type": "string",
                        "format": "string",
                        "maxLength": 300
                    },
                    "created_at": {
                        "schema": "Producer",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "schema": "Producer",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "schema": "Producer",
                        "description": "deleted_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "schema": "Producer",
                        "description": "User",
                        "type": "object",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "image": {
                        "schema": "Producer",
                        "description": "File",
                        "type": "object",
                        "items": {
                            "$ref": "#/components/schemas/File"
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "description": "Entity User",
                "required": [
                    "id",
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "schema": "User",
                        "description": "id",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "name": {
                        "schema": "User",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "schema": "User",
                        "description": "email",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "email_verified_at": {
                        "schema": "User",
                        "description": "email_verified_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "password": {
                        "schema": "User",
                        "description": "password",
                        "type": "string",
                        "format": "password",
                        "maxLength": 255
                    },
                    "remember_token": {
                        "schema": "User",
                        "description": "remember_token",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "payment_account_id": {
                        "schema": "User",
                        "description": "payment_account_id",
                        "type": "integer"
                    },
                    "someone_id": {
                        "schema": "User",
                        "description": "someone_id",
                        "type": "integer"
                    },
                    "someone_type": {
                        "schema": "User",
                        "description": "someone_type",
                        "type": "string"
                    },
                    "last_login_at": {
                        "schema": "User",
                        "description": "last_login_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "last_login_ip": {
                        "schema": "User",
                        "description": "last_login_ip",
                        "type": "string",
                        "format": "string"
                    },
                    "created_at": {
                        "schema": "User",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "schema": "User",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://roundapp.local/oauth/authorize",
                        "tokenUrl": "https://roundapp.local/oauth/token",
                        "refreshUrl": "https://roundapp.local/token/refresh",
                        "scopes": [],
                        "client_id": "1",
                        "secret": "yo4kJNkiGGIiU0YqZ5jP1ucCArLM11FC7z5QdEFY"
                    }
                }
            }
        }
    },
...
}